Applicate modifiche al controller JWT: da /autenticazione a /login Modificata l'entità categoria.

+++ jwt-youtube-master/src/main/java/it/sincrono/ecommerce/repository/CategoriaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import it.sincrono.ecommerce.entity.Categoria;
 
 @Repository
-public interface CategoriaRepository extends CrudRepository<Categoria, Long>{
+public interface CategoriaRepository extends CrudRepository<Categoria, Integer>{
 
 }
